Adjust the optional schemas generation
