+++ vaadin-connect-maven-plugin/src/main/java/com/vaadin/connect/plugin/generator/VaadinConnectTsGenerator.java
@@ -25,7 +25,6 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -57,7 +56,6 @@ import io.swagger.v3.oas.models.media.ComposedSchema;
 import io.swagger.v3.oas.models.media.Content;
 import io.swagger.v3.oas.models.media.MapSchema;
 import io.swagger.v3.oas.models.media.MediaType;
-import io.swagger.v3.oas.models.media.ObjectSchema;
 import io.swagger.v3.oas.models.media.Schema;
 import io.swagger.v3.oas.models.parameters.RequestBody;
 import io.swagger.v3.oas.models.tags.Tag;
@@ -794,21 +792,13 @@ public class VaadinConnectTsGenerator extends AbstractTypeScriptClientCodegen {
     } else if (schema.getAdditionalProperties() != null) {
       Schema inner = (Schema) schema.getAdditionalProperties();
       return String.format("{ [key: string]: %s; }", getTypeDeclaration(inner));
-    } else if (isOptionalSchema(schema)) {
+    } else if (OpenApiObjectGenerator.isOptionalSchema(schema)) {
       return String.format("%s | null",
           getTypeDeclaration(((Map<String, Schema>) schema.getProperties())
               .values().iterator().next()));
     } else {
       return super.getTypeDeclaration(schema);
     }
-  }
-
-  private boolean isOptionalSchema(Schema schema) {
-    return schema instanceof ObjectSchema
-        && Optional.ofNullable(schema.getRequired()).map(List::isEmpty)
-            .orElse(true)
-        && Optional.ofNullable(schema.getProperties())
-            .filter(properties -> properties.size() == 1).isPresent();
   }
 
   private Schema getRequestBodySchema(RequestBody body) {
