+++ vaadin-connect-maven-plugin/src/test/java/com/vaadin/connect/plugin/generator/services/AbstractServiceGenerationTest.java
@@ -348,10 +348,6 @@ public abstract class AbstractServiceGenerationTest {
   }
 
   private void assertSchema(Schema actualSchema, Class<?> expectedSchemaClass) {
-    if (expectedSchemaClass.isAssignableFrom(Optional.class)) {
-      //FIXME:
-      return;
-    }
     if (actualSchema.get$ref() != null) {
       assertNull(actualSchema.getProperties());
       schemaReferences.add(actualSchema.get$ref());
@@ -450,8 +446,7 @@ public abstract class AbstractServiceGenerationTest {
     if (properties.isEmpty()) {
       assertNull(schema.getRequired());
     } else {
-      //FIXME:
-//      assertEquals(new HashSet<>(schema.getRequired()), properties.keySet());
+      assertEquals(new HashSet<>(schema.getRequired()), properties.keySet());
     }
   }
