+++ vaadin-connect-maven-plugin/src/test/resources/com/vaadin/connect/plugin/generator/services/json/expected-model-com.vaadin.connect.plugin.generator.services.json.JsonTestService.User.ts
@@ -1,7 +1,7 @@
 import Role from './Role';
 
 export default interface User {
-  a: string | null;
+  optionalField: string | null;
   name: string;
   password: string;
   roles: { [key: string]: Role; };
