+++ vaadin-connect-maven-plugin/src/test/resources/com/vaadin/connect/plugin/generator/services/model/expected-model-com.vaadin.connect.plugin.generator.services.model.ModelService.Account.ts
@@ -9,7 +9,6 @@ export default interface Account {
    */
   groups: Group[];
   modelFromDifferentPackage: ModelFromDifferentPackage;
-  optionalField: string | null;
   /**
    * Javadoc for username.
    */
